backend/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
pydantic[email]
python-telegram-bot
python-multipart